# specs/002-coach-availability-impersonation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PracticePerfect — Coach Invitations, Availability & Impersonation
  version: 1.3.0
  description: |
    Contract for feature 002-coach-availability-impersonation. It is version **1.3.0** of the same
    `/api/v1` contract feature 001 defined at 1.2.0, and it is **purely additive**: every operation
    below is new, and the only changes to existing response models are new optional fields
    (`CurrentUser.impersonation`, `CurrentUser.impersonation_ended`,
    `TrainerPlayerSummary.availability`, `TrainerPlayerSummary.availability_updated_at`). No
    operation is removed, and no existing field changes type or meaning, so the frontend compiled
    against 1.2.0 keeps working until it chooses to read the new fields (research.md R2-22).

    This file documents **only what 1.3.0 adds**. Operations defined in
    `specs/001-user-roles-admin/contracts/openapi.yaml` are unchanged and not repeated here, with
    the single exception of the two schema extensions above, which are given in full so the
    frontend types have one authoritative source.

    Three groups of operations:

    * **Coach invitations** (`coach-invitations` tag, plus `/trainer/coach-invitations` and
      `/trainer/coaches`) — a trainer invites a coach, tracks the invitation, and manages the
      roster; the invited person previews, registers, or accepts. US-01.08, FR-001 – FR-023.
    * **Availability** (`availability` tag) — a coach's own week, a player profile's week, and the
...

...

...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /me/availability:
    get:
      tags: [availability]
      summary: Read the signed-in coach's own stated week
      operationId: getOwnAvailability
      x-requirements: [FR-024, FR-032, FR-033, FR-035]
      description: |
        Coach only. An account that has never stated times returns `slots: []` with
        `updated_at: null` — "no times set", which FR-035 requires be distinguishable from
        unavailable and from a week that was deliberately cleared (`slots: []` with a non-null
        `updated_at`).
      responses:
        '200':
          description: The coach's stated week.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilityWeek' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }

    put:
      tags: [availability]
      summary: Replace the signed-in coach's stated week
      operationId: replaceOwnAvailability
      x-requirements: [FR-024, FR-026, FR-027, FR-028, FR-029, FR-032, FR-033]
      description: |
        Replaces the whole week in one transaction: the submitted set is validated in full before
        any row is touched, so a refused save leaves the previous week exactly as it was and no week
        is ever observable half-saved (FR-027, FR-029). Two concurrent saves resolve as
        last-complete-save-wins; there is no version token to send.

        `422` carries the offending **day** and the reason — overlapping ranges, a start at or after
        its end, a range past the end of the day, or more than six ranges in a day.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AvailabilityWeekUpdate' }
      responses:
        '200':
          description: The stored week, as saved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilityWeek' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }

    delete:
      tags: [availability]
      summary: Clear the signed-in coach's stated week
      operationId: clearOwnAvailability
      x-requirements: [FR-030, FR-032]
      description: |
        Equivalent to saving an empty week, and kept as its own operation because "I have no times
        to offer" is a statement a person makes deliberately. `updated_at` is stamped, so the week
        reads as cleared-on-a-date rather than never-stated.
      responses:
        '204': { description: The week is cleared. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /me/players/{profile_id}/availability:
    get:
      tags: [availability]
      summary: Read a player profile's stated week
      operationId: getProfileAvailability
      x-requirements: [FR-025, FR-032, FR-033, FR-035]
      description: |
        Nested under the family resource, so ownership is the check that resource already performs:
        a parent reaches their own profile and each child's; a signed-in child reaches their own and
        nothing else. An unreachable profile — another account's, or a sibling's when the caller is
        a child — is a 404, not a 403, exactly as feature 001's family routes behave.
      parameters:
        - $ref: '#/components/parameters/PlayerProfileId'
      responses:
        '200':
          description: That profile's stated week.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilityWeek' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    put:
      tags: [availability]
      summary: Replace a player profile's stated week
      operationId: replaceProfileAvailability
      x-requirements: [FR-025, FR-026, FR-027, FR-028, FR-029, FR-032, FR-033]
      description: |
        Same whole-week semantics and same validation as `PUT /me/availability`. A parent may save
        for their own profile and for any child's; a signed-in child may save only their own
        (FR-033) — the parent retains authority to revise what a child stated.
      parameters:
        - $ref: '#/components/parameters/PlayerProfileId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AvailabilityWeekUpdate' }
      responses:
        '200':
          description: The stored week, as saved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilityWeek' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }

    delete:
      tags: [availability]
      summary: Clear a player profile's stated week
      operationId: clearProfileAvailability
      x-requirements: [FR-030, FR-032, FR-033]
      parameters:
        - $ref: '#/components/parameters/PlayerProfileId'
      responses:
        '204': { description: The week is cleared. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /admin/impersonations:
    post:
      tags: [impersonation]
      summary: Start impersonating a user
      operationId: startImpersonation
      x-requirements: [FR-040, FR-041, FR-042, FR-048, FR-049, FR-051, FR-052]
      description: |
        Super Admin only, enforced on the request. No new session is created: the caller's existing
        session begins resolving to the target account, and the impersonated person's own sessions
        are untouched (FR-049).

        Refused with `422 impersonation_not_permitted` when the target is another Super Admin, the
        caller themselves, or an account whose personal information has been erased (FR-042). An
        Inactive target is permitted and is labelled in the response.

        If the caller already holds an open impersonation it is closed as `superseded` in the same
        transaction (FR-048).

        Because the caller's own session now resolves to the target, this endpoint is unreachable
        *while* impersonating — the effective user is not a Super Admin — which is how FR-047's
        no-nested-impersonation rule is structural rather than hand-written (research.md R2-15).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImpersonationCreate' }
      responses:
        '201':
          description: The impersonation is live; the caller's next request is the target's.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Impersonation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ActionNotPermitted' }

    get:
      tags: [impersonation]
      summary: Read the impersonation history
      operationId: listImpersonations
      x-requirements: [FR-053, FR-054, FR-055, FR-056]
      description: |
        Super Admin only; nothing is returned to any other role (FR-056). Every impersonation the
        platform has ever permitted, including any still in progress (`ended_at: null`), newest
        first. `duration_seconds` is computed, not stored, and is `null` while in progress.

        Entries survive the erasure of the impersonated account: once erasure has removed the
        personal details, the row still names the account by identifier (FR-055).
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: admin_user_id
          in: query
          required: false
          description: Only impersonations started by this Super Admin.
          schema: { type: string, format: uuid }
        - name: target_user_id
          in: query
          required: false
          description: Only impersonations of this account.
          schema: { type: string, format: uuid }
        - name: started_from
          in: query
          required: false
          description: Only impersonations that started at or after this instant (UTC).
          schema: { type: string, format: date-time }
        - name: started_to
          in: query
          required: false
          description: Only impersonations that started at or before this instant (UTC).
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: One page of history entries.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImpersonationPage' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /admin/impersonations/current:
    delete:
      tags: [impersonation]
      summary: Leave the impersonation in progress
      operationId: endImpersonation
      x-requirements: [FR-045, FR-046, FR-051]
      description: |
        **The one operation in the contract that authorizes on the caller's real identity rather
        than their effective one** (research.md R2-15). While impersonating a Trainer the effective
        user is a Trainer, so a Super Admin role gate here would lock the admin inside the
        impersonation until the one-hour ceiling. The gate is instead: an Active Super Admin session
        holding an open impersonation.

        Returns the admin to their own view with their own permissions and no re-authentication. The
        impersonation is closed as `exited`.

        `404` when the caller holds no open impersonation — including the case where it has just
        ended on the deadline, which the caller learns from `GET /auth/session`.
      responses:
        '204': { description: The impersonation has ended; the caller is themselves again. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ==========================================================================
  # Extended existing operation (schema additions only)
  # ==========================================================================
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pp_session
      description: |
        Unchanged from feature 001. While an impersonation is live the same cookie resolves to the
        impersonated account; no second credential is issued (research.md R2-14).
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
    PlayerProfileId:
      name: profile_id
      in: path
      required: true
      description: |
        A player profile. Under `/me/players` it must be on the caller's own account; under
        `/trainer/players` it must hold an Active association with the caller. Unreachable profiles
        are 404, as in feature 001.
      schema: { type: string, format: uuid }
  responses:
    Unauthenticated:
      description: No usable session cookie.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The caller's role or ownership does not permit this action.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such resource is reachable by this caller.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ActionNotPermitted:
      description: Understood, well-formed, and refused by a domain rule.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: One or more fields are invalid; each offending field is identified.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorResponse' }
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: true
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    ValidationErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [message, fields]
          properties:
            message: { type: string }
            fields:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }

    # --- Coach invitations ---
...
    AvailabilitySlot:
      type: object
      additionalProperties: false
      required: [day_of_week, start_minute, end_minute]
      description: |
        One range on one day. Minutes from midnight on a 15-minute grid (research.md R2-08):
        `17:00–20:00` is `{day_of_week: 0, start_minute: 1020, end_minute: 1200}`.
      properties:
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          description: 0 = Monday … 6 = Sunday.
        start_minute:
          type: integer
          minimum: 0
          maximum: 1425
          multipleOf: 15
        end_minute:
          type: integer
          minimum: 15
          maximum: 1440
          multipleOf: 15
          description: May be 1440 (midnight). Must be greater than `start_minute`.
    AvailabilityWeek:
      type: object
      additionalProperties: false
      required: [slots, updated_at]
      properties:
        slots:
          type: array
          maxItems: 42
          description: |
            Ordered by `(day_of_week, start_minute)`. At most six per day (FR-028), so at most 42.
          items: { $ref: '#/components/schemas/AvailabilitySlot' }
        updated_at:
          type: [string, 'null']
          format: date-time
          description: |
            `null` means never stated. A non-null value with an empty `slots` means deliberately
            cleared. Neither means "unavailable" (FR-035).
    AvailabilityWeekUpdate:
      type: object
      additionalProperties: false
      required: [slots]
      description: |
        The complete week. This is a replacement, not a patch: whatever is absent is removed
        (FR-029). An empty array is equivalent to `DELETE`.
      properties:
        slots:
          type: array
          maxItems: 42
          items: { $ref: '#/components/schemas/AvailabilitySlot' }

    # --- Impersonation ---
    ImpersonationCreate:
      type: object
      additionalProperties: false
      required: [user_id]
      properties:
        user_id:
          type: string
          format: uuid
          description: The account to impersonate. Not a Super Admin, not the caller, not erased.
    ImpersonationEndReason:
      type: string
      enum:
        [exited, timed_out, signed_out, superseded, target_deactivated, target_erased,
         admin_deactivated]
    Impersonation:
      type: object
      additionalProperties: false
      required:
        [id, admin, target, target_status_at_start, started_at, expires_at, ended_at, end_reason,
         duration_seconds]
      properties:
        id: { type: string, format: uuid }
        admin: { $ref: '#/components/schemas/ImpersonationParticipant' }
        target: { $ref: '#/components/schemas/ImpersonationParticipant' }
        target_status_at_start:
          type: string
          enum: [active, inactive]
          description: |
            What makes FR-050's "leaves Active status" computable, and what the client labels an
            Inactive impersonation with (FR-042, research.md R2-19).
        started_at: { type: string, format: date-time }
        expires_at:
          type: string
          format: date-time
          description: The one-hour ceiling. Not extended by activity (FR-046).
        ended_at: { type: [string, 'null'], format: date-time }
        end_reason:
          oneOf:
            - $ref: '#/components/schemas/ImpersonationEndReason'
            - type: 'null'
        duration_seconds:
          type: [integer, 'null']
          description: Computed from the two timestamps, never stored. `null` while in progress.
    ImpersonationParticipant:
      type: object
      additionalProperties: false
      required: [user_id, display_name, role]
      description: |
        Named by identifier first: after the impersonated account is erased, `display_name` is the
        anonymized name feature 001's erasure leaves behind, and the entry still stands (FR-055).
      properties:
        user_id: { type: string, format: uuid }
        display_name: { type: string }
        role: { type: string, enum: [super_admin, trainer, coach, player_parent] }
    ImpersonationPage:
      type: object
      additionalProperties: false
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Impersonation' }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }

    # --- Extended: CurrentUser (1.2.0 fields elided; the two additions are normative here) ---
...